Prefer `Device_Name` to `Device_Code_Name` for browser.DeviceName.
Before, it was only looking at Device_Code_Name.  But Device_Name is a
separate field that sometimes has a different value than
Device_Code_Name.  This change makes it so we prefer the former to the
latter.

This is an API-breaking change, but I think it's the right thing to do
based on my (admittedly scant) understanding of these ini fields.